Re-write TimeInterpreter and use unsafeExtendSafeZone where needed
TimeInterpreter was re-written to have more powerful combinators and
useful logging.

To fix integration tests, unsafeExtendSafeZone was used in the following
places:

1. getTxExpiry for setting TTL in the future
2. mkApiTransaction for showing TTL in the future
3. mkShelleyWallet for delegation status changes
4. combineDbAndLsqData for future pool retirement dates

This allows dates, slots and epochs to be converted between even in the
Shelley era of Byron;Shelley;Allegra;Mary — but the results could also
be **wrong**.

It seems the slotting parameters won't change in the coming eras. It
would still be nice to replace unsafeExtendSafeZone with something else,
in the future.