Revise worker error handling
With MonadUnliftIO it becomes easier to write polymorphic
functions (MonadIO is not enough when there are callbacks).

So we can make the worker thread registry polymorphic, and run it in
IO or ExceptT e IO.

That removes the need for using the unsafeRunExceptT function.

Although the register function can now work with checked exception
types, it is also modified to be aware of IO exceptions which may be
thrown during setup. These other exceptions are passed up to the
caller.

The worker registry changes also fix a race condition where multiple
worker threads could be started for the for the same key, before the
worker resource is acquired.