Minimize the central monadic expression of `makeChange`.
There are two sets of values in this function that are completely
determined by the function's arguments.

Before this change, some were listed in the `where` clause, but others
were listed in the `let` clause of the function body.

This change simply consolidates them into a single place.

As a result, the function's main monadic expression is minimized.